python/requirements.txt
openai-whisper==20231117
transformers==4.35.2
torch==2.2.1
sentencepiece==0.2.0
accelerate==0.24.1
tiktoken==0.7.0
numpy<2
huggingface-hub==0.23.2
hf_xet>=0.5.0

